db_init/requirements.txt
joblib
numpy
pymysql
python-dotenv
scikit-learn